requirements.txt
asgi-redis==1.1.0
asgiref==3.5.2
attrs==21.4.0
autobahn==22.2.2
Automat==20.2.0
Babel==2.9.1
Brotli==1.1.0
cement==2.8.2
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.3
colorthief==0.2.1
constantly==15.1.0
cryptography==36.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.12
django-classy-tags==4.0.0
django-cms==3.11.3
django-crispy-forms==1.14.0
django-decouple==2.1
django-extra-views==0.13.0
django-filter==21.1
django-formtools==2.4.1
django-haystack==3.1.1
django-jazzmin==2.6.0
django-material-icons==0.1.2
django-phonenumber-field==5.2.0
django-rest-passwordreset==1.3.0
django-sekizai==4.1.0
django-storages==1.12.3
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangocms-admin-style==3.2.4
djangorestframework==3.13.1
docutils==0.15.2
factory-boy==3.2.1
Faker==13.0.0
future==0.16.0
gunicorn==22.0.0
h11==0.14.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jmespath==0.10.0
joblib==1.1.0
msgpack-python==0.5.6
packaging==23.1
Paperboy==1.0.1
pathspec==0.9.0
phonenumbers==8.12.43
Pillow==9.0.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
purl==1.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.1.10
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==22.0.0
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.19.2
pytz==2021.3
redis==2.10.6
regex==2022.1.18
requests==2.26.0
semantic-version==2.8.5
service-identity==21.1.0
simplejson==3.17.5
six==1.14.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
termcolor==1.1.0
text-unidecode==1.3
tqdm==4.62.3
twilio==7.6.0
txaio==22.2.1
typing_extensions==4.1.1
tzdata==2023.3
urllib3==1.26.9
uvicorn==0.30.3
wcwidth==0.1.9
whitenoise==6.0.0
